app/requirements.txt
fastapi
uvicorn
pillow
neo4j